pyproject.toml
[tool.poetry]
name = "selenium-web-scraping"
version = "0.1.0"
description = ""
authors = ["vova2092 <vovai2092@yandex.ru>"]
readme = "README.md"
packages = [{include = "selenium_web_scraping"}]

[tool.poetry.dependencies]
python = "^3.11"
bs4 = "^0.0.1"
selenium = "^4.9.1"
lxml = "^4.9.2"
python-dotenv = "^1.0.0"
pydantic = "^1.10.7"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
